+++ src/Sentry/Platforms/Cocoa/CocoaScopeObserver.cs
@@ -89,7 +89,7 @@ internal sealed class CocoaScopeObserver : IScopeObserver
         }
     }
 
-    public void SetUser(User? user)
+    public void SetUser(SentryUser? user)
     {
         try
         {