+++ test/Sentry.Tests/Protocol/SentryEventTests.verify.cs
@@ -11,7 +11,7 @@ public partial class SentryEventTests
         var id = Guid.Parse("4b780f4c-ec03-42a7-8ef8-a41c9d5621f8");
         var sut = new SentryEvent(ex, timestamp, id)
         {
-            User = new User { Id = "user-id" },
+            User = new SentryUser { Id = "user-id" },
             Request = new Request { Method = "POST" },
             Contexts = new Contexts
             {